# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/files": {
      "get": {
        "operationId": "FilesController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "文件元数据数组",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "mimeType": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "createdAt": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "mimeType",
                          "size",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token（UNAUTHORIZED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误（INTERNAL_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "文件列表",
        "tags": [
          "files"
        ]
      },
      "post": {
        "operationId": "FilesController_upload_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "文件二进制内容"
                  },
                  "name": {
                    "type": "string",
                    "description": "文件名"
                  },
                  "mimeType": {
                    "type": "string",
                    "description": "MIME 类型"
                  }
                },
                "required": [
                  "file",
                  "name",
                  "mimeType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传后的文件元数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "mimeType": {
                          "type": "string"
                        },
                        "size": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "mimeType",
                        "size",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token（UNAUTHORIZED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误（INTERNAL_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "501": {
            "description": "能力尚未实现（NOT_IMPLEMENTED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "上传文件（multipart/form-data）",
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{id}": {
      "get": {
        "operationId": "FilesController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "文件 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文件元数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "mimeType": {
                          "type": "string"
                        },
                        "size": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "mimeType",
                        "size",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token（UNAUTHORIZED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "资源不存在（NOT_FOUND）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误（INTERNAL_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取文件元数据",
        "tags": [
          "files"
        ]
      },
      "delete": {
        "operationId": "FilesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "文件 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功（响应体为 {}）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token（UNAUTHORIZED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误（INTERNAL_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "501": {
            "description": "能力尚未实现（NOT_IMPLEMENTED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除文件",
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "info": {
    "title": "Growth OS API",
    "description": "Growth OS 后端服务 API 文档（自动生成）。\n\n**响应信封**：成功响应统一为 `{ data: T }`（204 除外）；失败统一为 `ApiErrorEnvelope { code, message, details? }`，`code` 为机器可读错误码，前端按码分支处理。\n\n**鉴权**：除 health 域外，所有端点均需 Bearer JWT（Supabase Auth access token）——点击右上角 Authorize 输入 token。",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}